Refactoring everything into smaller files.
