+++ src/rubik/paths/__init__.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env python3
+
+if __name__ == "__main__":
+    pass
