+++ src/rubik/shapes/__init__.py
@@ -1,0 +1,10 @@
+from .tile import Tile
+from .domino import Domino
+from .strip import Strip
+from .square import Square
+from .grid import Grid
+from .sheet import Sheet
+from .volume import Volume
+from .cube import Cube
+from .triangle import Triangle
+from .tetrahedron import Tetrahedron
