+++ src/rubik/shapes/cube.py
@@ -1,0 +1,45 @@
+#!/usr/bin/env python3
+
+from .shape import Shape
+from common.logger import log
+from copy import deepcopy
+from rubik.paths.moves import Move
+import numpy as np
+
+
+class Cube(Shape):
+    """
+    A 3x3x3 set of square tiles. (The classic Rubik's cube).
+
+           ____ ____ ____
+         /____/____/____/|
+       /____/____/____/| |
+     /____/____/____/| |/|
+    |    |    |    | |/| |
+    |    |    |    |/| | |
+     ---- ---- ----  | |/|
+    |    |    |    | |/| |
+    |    |    |    |/| | |
+     ---- ---- ----  | |/
+    |    |    |    | |/
+    |    |    |    |/
+     ---- ---- ----
+
+    """
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        for colour in self.colours:
+            self.faces.append([[colour.value for i in range(3)] for j in range(3)])
+            if len(self.faces) == 6:
+                break
+        assert len(self.faces) == 6, f"A {type(self).__name__} must have only 6 faces."
+        for face in self.faces:
+            assert np.shape(face) == (3, 3), f"A {type(self).__name__} face must only contain 9 tiles."
+
+    def moves(self):
+        raise NotImplementedError
+
+
+if __name__ == "__main__":
+    pass
